refactor: Cookie処理を汎用化してユーティリティ関数に移行
- utils/cookie.tsにServer Components用のCookie処理を汎用化
- getCookieHeaderForAPI: API呼び出し用Cookieヘッダー取得関数
- createAPIHeaders: API用headersオブジェクト作成関数
- contact.tsで新しいユーティリティ関数を使用
